requirements.txt
PyYAML==3.11
ply==3.11
llvmlite==0.30.0